all required functions compete

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: pthorell <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/09 14:17:37 by pthorell          #+#    #+#             */
-/*   Updated: 2018/07/10 19:27:32 by pthorell         ###   ########.fr       */
+/*   Updated: 2018/07/11 15:53:51 by pthorell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	int i;
+	int	i;
 
 	i = 0;
 	if (dst <= src)
 	{
-		while (i < len)
+		while (i < (int)len)
 		{
 			((char*)dst)[i] = ((char*)src)[i];
 			++i;
