+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: pthorell <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/09 17:12:59 by pthorell          #+#    #+#             */
-/*   Updated: 2018/07/10 20:45:56 by pthorell         ###   ########.fr       */
+/*   Updated: 2018/07/11 15:50:09 by pthorell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
 	int		max;
-	int		i;
+	size_t	i;
 	int		j;
-	int		len;
+	size_t	len;
 	int		c;
 
 	len = ft_strlen(dst);
