+++ ft_strnstr.c
@@ -6,16 +6,16 @@
 /*   By: pthorell <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/09 19:30:08 by pthorell          #+#    #+#             */
-/*   Updated: 2018/07/10 21:56:44 by pthorell         ###   ########.fr       */
+/*   Updated: 2018/07/11 16:04:06 by pthorell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
-{	
-	int s_len;
-	int	i;
+{
+	int		s_len;
+	size_t	i;
 
 	if (len == 0)
 		return (NULL);
@@ -23,7 +23,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 		return ((char*)haystack);
 	i = 0;
 	s_len = ft_strlen(needle);
-	while (haystack[i] && i <= (len - s_len))
+	while (haystack[i] && (int)i <= ((int)len - s_len))
 	{
 		if (!ft_strncmp(haystack + i, needle, s_len))
 			return ((char*)(haystack + i));
